🚚 api: move deneutralize-challenge route to src/certification/evaluation
