--- api/lib/application/certifications/index.js
@@ -1,42 +1,0 @@
-import Joi from 'joi';
-
-import { securityPreHandlers } from '../../../src/shared/application/security-pre-handlers.js';
-import { identifiersType } from '../../../src/shared/domain/types/identifiers-type.js';
-import { certificationController } from './certification-controller.js';
-
-const register = async function (server) {
-  server.route([
-    {
-      method: 'POST',
-      path: '/api/admin/certification/deneutralize-challenge',
-      config: {
-        validate: {
-          payload: Joi.object({
-            data: {
-              attributes: {
-                certificationCourseId: identifiersType.certificationCourseId,
-                challengeRecId: Joi.string().required(),
-              },
-            },
-          }),
-        },
-        pre: [
-          {
-            method: (request, h) =>
-              securityPreHandlers.hasAtLeastOneAccessOf([
-                securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-                securityPreHandlers.checkAdminMemberHasRoleCertif,
-                securityPreHandlers.checkAdminMemberHasRoleSupport,
-              ])(request, h),
-            assign: 'hasAuthorizationToAccessAdminScope',
-          },
-        ],
-        handler: certificationController.deneutralizeChallenge,
-        tags: ['api'],
-      },
-    },
-  ]);
-};
-
-const name = 'certifications-api';
-export { name, register };
