+++ api/lib/domain/events/handle-certification-rescoring.js
@@ -1,3 +1,4 @@
+import { ChallengeDeneutralized } from '../../../src/certification/evaluation/domain/events/ChallengeDeneutralized.js';
 import { ChallengeNeutralized } from '../../../src/certification/evaluation/domain/events/ChallengeNeutralized.js';
 import { services } from '../../../src/certification/evaluation/domain/services/index.js';
 import { AssessmentResultFactory } from '../../../src/certification/scoring/domain/models/factories/AssessmentResultFactory.js';
@@ -10,7 +11,6 @@ import { CertificationCourseUnrejected } from './CertificationCourseUnrejected.j
 import { CertificationJuryDone } from './CertificationJuryDone.js';
 import CertificationRescoredByScript from './CertificationRescoredByScript.js';
 import { CertificationRescoringCompleted } from './CertificationRescoringCompleted.js';
-import { ChallengeDeneutralized } from './ChallengeDeneutralized.js';
 import { checkEventTypes } from './check-event-types.js';
 
 const eventTypes = [
