+++ api/lib/routes.js
@@ -4,7 +4,6 @@ import * as campaignParticipations from './application/campaign-participations/i
 import * as certificationCenterInvitations from './application/certification-center-invitations/index.js';
 import * as certificationCenterMemberships from './application/certification-center-memberships/index.js';
 import * as certificationCenters from './application/certification-centers/index.js';
-import * as certifications from './application/certifications/index.js';
 import * as frameworks from './application/frameworks/index.js';
 import * as memberships from './application/memberships/index.js';
 import * as organizations from './application/organizations/index.js';
@@ -22,7 +21,6 @@ const routes = [
   certificationCenters,
   certificationCenterInvitations,
   certificationCenterMemberships,
-  certifications,
   healthcheck,
   memberships,
   organizations,
