+++ api/src/certification/evaluation/application/certification-controller.js
@@ -14,8 +14,22 @@ const neutralizeChallenge = async function (request, h, dependencies = { events 
   return h.response().code(204);
 };
 
+const deneutralizeChallenge = async function (request, h, dependencies = { events }) {
+  const challengeRecId = request.payload.data.attributes.challengeRecId;
+  const certificationCourseId = request.payload.data.attributes.certificationCourseId;
+  const juryId = request.auth.credentials.userId;
+  const event = await usecases.deneutralizeChallenge({
+    challengeRecId,
+    certificationCourseId,
+    juryId,
+  });
+  await dependencies.events.eventDispatcher.dispatch(event);
+  return h.response().code(204);
+};
+
 const certificationController = {
   neutralizeChallenge,
+  deneutralizeChallenge,
 };
 
 export { certificationController };
