+++ api/src/certification/evaluation/domain/usecases/deneutralize-challenge.js
@@ -1,10 +1,10 @@
 import { ChallengeDeneutralized } from '../events/ChallengeDeneutralized.js';
 
 const deneutralizeChallenge = async function ({
-  certificationAssessmentRepository,
   certificationCourseId,
   challengeRecId,
   juryId,
+  certificationAssessmentRepository,
 }) {
   const certificationAssessment = await certificationAssessmentRepository.getByCertificationCourseId({
     certificationCourseId,
