+++ api/tests/certification/evaluation/unit/application/certification-route_test.js
@@ -83,4 +83,84 @@ describe('Certification | Evaluation | Unit | Application | Routes | certificati
       expect(response.statusCode).to.equal(400);
     });
   });
+
+  context('POST /api/admin/certification/deneutralize-challenge', function () {
+    it('return forbidden access if user has METIER role', async function () {
+      // given
+      sinon
+        .stub(securityPreHandlers, 'hasAtLeastOneAccessOf')
+        .withArgs([
+          securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
+          securityPreHandlers.checkAdminMemberHasRoleCertif,
+          securityPreHandlers.checkAdminMemberHasRoleSupport,
+        ])
+        .callsFake(
+          () => (request, h) =>
+            h
+              .response({ errors: new Error('forbidden') })
+              .code(403)
+              .takeover(),
+        );
+
+      const httpTestServer = new HttpTestServer();
+      await httpTestServer.register(moduleUnderTest);
+
+      // when
+      const response = await httpTestServer.request('POST', '/api/admin/certification/deneutralize-challenge', {
+        data: {
+          attributes: {
+            certificationCourseId: 1,
+            challengeRecId: 'rec43mpMIR5dUzdjh',
+          },
+        },
+      });
+
+      // then
+      expect(response.statusCode).to.equal(403);
+    });
+
+    it('checks that a valid certification-course id is given', async function () {
+      // given
+      sinon.stub(certificationController, 'deneutralizeChallenge').returns('ok');
+      sinon.stub(securityPreHandlers, 'hasAtLeastOneAccessOf').returns(() => true);
+      const httpTestServer = new HttpTestServer();
+      await httpTestServer.register(moduleUnderTest);
+      const payload = {
+        data: {
+          attributes: {
+            certificationCourseId: 'invalide',
+            challengeRecId: 'rec43mpMIR5dUzdjh',
+          },
+        },
+      };
+
+      // when
+      const response = await httpTestServer.request('POST', '/api/admin/certification/deneutralize-challenge', payload);
+
+      // then
+      expect(response.statusCode).to.equal(400);
+    });
+
+    it('checks that a challenge recId is given', async function () {
+      // given
+      sinon.stub(certificationController, 'deneutralizeChallenge').returns('ok');
+      sinon.stub(securityPreHandlers, 'hasAtLeastOneAccessOf').returns(() => true);
+      const httpTestServer = new HttpTestServer();
+      await httpTestServer.register(moduleUnderTest);
+      const payload = {
+        data: {
+          attributes: {
+            certificationCourseId: 1,
+            challengeRecId: null,
+          },
+        },
+      };
+
+      // when
+      const response = await httpTestServer.request('POST', '/api/admin/certification/deneutralize-challenge', payload);
+
+      // then
+      expect(response.statusCode).to.equal(400);
+    });
+  });
 });
