+++ api/tests/certification/evaluation/unit/domain/services/scoring/scoring-v2_test.js
@@ -1,6 +1,6 @@
 import _ from 'lodash';
 
-import { ChallengeDeneutralized } from '../../../../../../../lib/domain/events/ChallengeDeneutralized.js';
+import { ChallengeDeneutralized } from '../../../../../../../src/certification/evaluation/domain/events/ChallengeDeneutralized.js';
 import { ChallengeNeutralized } from '../../../../../../../src/certification/evaluation/domain/events/ChallengeNeutralized.js';
 import {
   calculateCertificationAssessmentScore,
