+++ api/tests/certification/evaluation/unit/domain/usecases/deneutralize-challenge_test.js
@@ -1,7 +1,7 @@
-import { ChallengeDeneutralized } from '../../../../lib/domain/events/ChallengeDeneutralized.js';
-import { deneutralizeChallenge } from '../../../../lib/domain/usecases/deneutralize-challenge.js';
-import { CertificationAssessment } from '../../../../src/certification/session-management/domain/models/CertificationAssessment.js';
-import { domainBuilder, expect, sinon } from '../../../test-helper.js';
+import { ChallengeDeneutralized } from '../../../../../../src/certification/evaluation/domain/events/ChallengeDeneutralized.js';
+import { deneutralizeChallenge } from '../../../../../../src/certification/evaluation/domain/usecases/deneutralize-challenge.js';
+import { CertificationAssessment } from '../../../../../../src/certification/session-management/domain/models/CertificationAssessment.js';
+import { domainBuilder, expect, sinon } from '../../../../../test-helper.js';
 
 describe('Unit | UseCase | deneutralize-challenge', function () {
   it('deneutralizes a challenge by its recId', async function () {
